refactor: code conventions III
