--- ee/packages/amplication-git-pull-service/src/constants.ts
@@ -1,13 +1,0 @@
-import { join } from "path";
-export const DOT_AMPLICATION_FOLDER = join(
-  process.cwd(),
-  "..",
-  "..",
-  ".amplication"
-);
-export const DEFAULT_GITHUB_PULL_FOLDER = join(
-  DOT_AMPLICATION_FOLDER,
-  "storage",
-  "github",
-  "pulls"
-);
