--- ee/packages/amplication-git-pull-service/src/errors.ts
@@ -1,11 +1,0 @@
-import * as common from "@nestjs/common";
-
-export class ForbiddenException extends common.ForbiddenException {
-  statusCode!: number;
-  message!: string;
-}
-
-export class NotFoundException extends common.NotFoundException {
-  statusCode!: number;
-  message!: string;
-}
