+++ ee/packages/amplication-git-pull-service/src/git-pull-event/git-pull-event.constants.ts
@@ -1,3 +1,18 @@
+import { join } from "path";
+
+export const DOT_AMPLICATION_FOLDER = join(
+  process.cwd(),
+  "..",
+  "..",
+  ".amplication"
+);
+export const DEFAULT_GITHUB_PULL_FOLDER = join(
+  DOT_AMPLICATION_FOLDER,
+  "storage",
+  "github",
+  "pulls"
+);
+
 // TODO: should be an enum
 // eslint-disable-next-line @typescript-eslint/naming-convention
 export const ErrorMessages = {
