--- ee/packages/amplication-git-pull-service/src/types.ts
@@ -1,3 +1,0 @@
-import { JsonValue } from "type-fest";
-
-export type InputJsonValue = Omit<JsonValue, "null">;
