--- ee/packages/amplication-git-pull-service/src/variables.ts
@@ -1,2 +1,0 @@
-export const CONFIG_SCOPE_GLOBAL = "global";
-export const CONFIG_SCOPE_LOCAL = "local";
